make ChartApplication
